requirements.txt
cohere>=4.47
voyageai>=0.1.2
faiss-cpu>=1.7.4
langchain>=0.1.11
openai>=1.12.0
datasets>=2.16.1
athina-client>=0.0.7
wandb>=0.16.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
bert-score>=0.3.13